requirements-dev.txt
Pillow==2.2.1
requests==2.0.1
tornado==3.1.1
simplejson==3.3.1
